scripts/requirements.txt
pdfplumber
pypdf2
pymupdf
python-dotenv
llama-parse
pytesseract
pdf2image
easyocr
docling